ATS_System/requirements.txt
streamlit
python-dotenv
openai
Pillow
pdf2image
opencv-python-headless